Create new NAND script
Create a new nand script that is running many tests comming
from mtd-utils such as nandbiterrs.

The script uses the name of the mtd partition to know on which
device the test must be performed (ie. Rootfs partition).
If the kernel is still using an old nand driver (pxa3xx), the
bitflip will not be correctly performed so "nandbiterrs" will
not be executed.

Currently, only the return of functions is tested. This script
must be improved.