nand: Check if NAND driver exists
Add a check if a NAND driver exists otherwise, return a
success because it is probably a missing configuration in
kernel defconfig.

For the moment, this case is handled as a "pass" but we will
update it to handle this case as a "skip". It is more significant
as the driver that we want to test is not available (so we can not
perform the test).

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>