nand: convert to use "custom" type
Convert the nand test to use "lava-test-case" directly
to be able to handle the "skip" use-case. It must use
the "custom" template not to use "--shell" option used in generic
template.

Create a "custom-template" script to define lava_pass/lava_fail
and lava_skip functions that handle the use of "lava-test-case".

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>