crypto: Improve it by looking at builtin driver
The script was testing only drivers compiled as modules.
In new kernel releases, the driver is compiled builtin. Because of that,
the crypto test ran correctly but as the modules' insertion failed,
the final crypto test failed too.

Improve this test by adding a check on builtin driver.
If the driver is not available builtin, we try to insert the module.
If it fails, the test fails.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>